review-feedback: 불필요한 추상화 제거
